2-6 enemy base
this should be the 'first working' of the enemies. if the methods are set up correctly, they start idling, then when they see luke within their sight area, they change to moving to a random position infront of luke. if they lose track of luke they go back to idle. if tis boba, then at a random point he will jetpack upwards and continue shooting at luke. boba can also ~1/3 or ~1/4 of the time fire a rocket at luke in a few vollys of rockets. haven't gotten the chance to fully test this as i haven't gotten help from chris to wire up a test luke for testing purposes, but it should be close to a proper first working.