Making changes
I added a new function attract() to use instead of gravity false. I also added rigid bodies to all the caves and valley. I changed some values in the navmesh bake so it is more precise. it still having problem with small objects and surfaces with slopes over around 60 degrees.